Checking vector add to be double instead of float

+++ SCMUlate/apps/vectorAddition/Codelets/vecAdd.cpp
@@ -1,7 +1,7 @@
 #include "vecAdd.hpp"
 #include <iomanip>
 
-#define numElements 2048*64/sizeof(int)
+#define numElements 2048*64/sizeof(double)
 
 IMPLEMENT_CODELET(vecAdd_2048L,
   // Obtaining the parameters
@@ -9,9 +9,9 @@ IMPLEMENT_CODELET(vecAdd_2048L,
   unsigned char *reg1 = args[0]; // Getting register 1
   unsigned char *reg2 = args[1]; // Getting register 2
   unsigned char *reg3 = args[2]; // Getting register 3
-  int *A = reinterpret_cast<int*>(reg2);
-  int *B = reinterpret_cast<int*>(reg3);
-  int *C = reinterpret_cast<int*>(reg1);
+  double *A = reinterpret_cast<double*>(reg2);
+  double *B = reinterpret_cast<double*>(reg3);
+  double *C = reinterpret_cast<double*>(reg1);
 
   for (uint64_t i = 0; i < numElements; i++)
     C[i] = A[i] + B[i];
